Renamed baseUri to baseMedia
baseMedia now corresponds to the serialized container which can be obtained
by calling GRiloMedia::serailze()

There is no unified way to create GrlMedia except by unserializing.
There is no proper way to start browsing from the middle of a tree in grilo.

Due to the way the declarative bindings are created, we will have to accept a
serialized media string and the model will have to unserialize it.